+++ src/components/Router/CrossDeviceMobileRouter.tsx
@@ -278,6 +278,7 @@ export default class CrossDeviceMobileRouter extends Component<
       const tokenPayload = getPayloadFromJWT(token)
       this.props.actions.setApplicantUuid(tokenPayload.app)
       this.props.actions.setClientUuid(tokenPayload.client_uuid)
+      this.props.actions.setIsTrial(tokenPayload.is_trial)
     }
 
     if (urls) {